1. fix redis cluster test 2. fix windows build
