+++ cmd/clone_test.go
